requirements.txt
aiogram==2.15
environs==9.3.4
Pillow==8.4.0